10968 fix sentry logging
